EN-2283 renamed newly created PageSettings* to SinglePageSettings* (to avoid name collisions with an existing PageSettings* about global page settings data)
