+++ sass/index.scss
@@ -42,7 +42,7 @@
 @import "pages/config/PageConfigGridRow";
 @import "pages/config/WidgetFrame";
 @import "pages/config/EmptyFrame";
-@import "pages/config/PageSettingsModal";
+@import "pages/config/SinglePageSettingsModal";
 @import "pages/list/PageListSearchTable";
 @import "users/common/UserStatus";
 @import "users/common/UserForm";
