+++ sass/pages/config/SinglePageSettingsModal.scss
@@ -1,4 +1,4 @@
-.PageSettingsModal {
+.SinglePageSettingsModal {
   .modal-title { // sass-lint:disable-line force-element-nesting class-name-format
     font-size: 16px;
   }
